backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
qdrant-client
sentence-transformers
torch
transformers
accelerate
numpy
google-generativeai